% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_down.R
\name{dr_down}
\alias{dr_down}
\title{Hardware report (experimental)}
\usage{
dr_down()
}
\description{
\code{dr_down} analyzes the R markdown capabilities of the hardware.
}
\examples{
\dontrun{
  indiedown::dr_down()
}
}
