% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indiedown-package.R
\docType{package}
\name{indiedown-package}
\alias{indiedown}
\alias{indiedown-package}
\title{indiedown: Individual R Markdown Templates}
\description{
Simplifies the generation of customized R Markdown PDF templates. A template may include an individual logo, typography, geometry or color scheme. The package provides a skeleton with detailed instructions for customizations. The skeleton can be modified by changing defaults in the 'YAML' header, by adding additional 'LaTeX' commands or by applying dynamic adjustments in R. Individual corporate design elements, such as a title page, can be added as R functions that produce 'LaTeX' code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://indiedown.cynkra.com/}
  \item \url{https://github.com/cynkra/indiedown}
  \item Report bugs at \url{https://github.com/cynkra/indiedown/issues}
}

}
\author{
\strong{Maintainer}: Christoph Sax \email{christoph.sax@gmail.com} (\href{https://orcid.org/0000-0002-7192-7044}{ORCID})

Authors:
\itemize{
  \item Kirill Müller (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})
  \item Angelica Becerra
  \item Frederik Aust
}

Other contributors:
\itemize{
  \item cynkra GmbH \email{mail@cynkra.com} (\href{https://ror.org/0335t7e62}{ROR}) [funder, copyright holder]
}

}
\keyword{internal}
