% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_indiedown_skeleton.R
\name{use_indiedown_skeleton}
\alias{use_indiedown_skeleton}
\title{Skeleton for a Customized RMarkdown Template}
\usage{
use_indiedown_skeleton(path = ".", overwrite = FALSE)
}
\description{
Skeleton for a Customized RMarkdown Template
}
\examples{
path <- file.path(tempdir(), "mydown")

# set up empty R Package 'mydown'
# with open = TRUE (default), it switches to the new project
usethis::create_package(path, open = FALSE)

# add indiedown assets
use_indiedown_skeleton(path, overwrite = TRUE)


}
