% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_indiedown_package.R
\name{create_indiedown_package}
\alias{create_indiedown_package}
\title{Create a Customized RMarkdown Template}
\usage{
create_indiedown_package(path = ".", overwrite = FALSE)
}
\description{
Create a Customized RMarkdown Template
}
\examples{

create_indiedown_package(file.path(tempdir(), "mydown"))
}
